move functions from quantum_manager.base to a new quantum_manager.utils module
